Extract Ignore helper class.
